Rodo fix

+++ app/routes/bucket.py
@@ -29,7 +29,7 @@ def get_data():
     buckets = Bucket.query.filter_by(user_id=user_id).all()  # Retrieve the buckets for the current user
 
     for obj in buckets:
-        data.append(obj.bucketfile_name)  # Append the 'bucketfile_name' attribute to the list
+        data.append((obj.bucketfile_name, obj.bucketfile_id))
 
     return jsonify(data)
 
@@ -42,7 +42,7 @@ def upload_to_s3():
     # Get the file name from the request
     file_name = request.form['filename']
     # Generate a unique file name
-    file_code = os.urandom(24).hex() + '.csv'
+    file_code = os.urandom(24).hex()
     # Get the user_id from the request
     user_id = current_user.get_id()
     print(user_id)
@@ -52,9 +52,16 @@ def upload_to_s3():
     if existing_file:
         return 'File with the same name already exists!', 409  # Return a conflict status code
 
-    # Check if the file has a CSV extension
-    if not file_name.lower().endswith('.csv'):
-        return 'Only CSV files are allowed!', 400  # Return a bad request status code
+    # Check if the file has a CSV or XLSX extension
+    if file_name.lower().endswith('.csv'):
+        file_extension = '.csv'
+    elif file_name.lower().endswith(('.xlsx', '.xls')):
+        file_extension = '.xlsx'
+    else:
+        return 'Only CSV and XLSX files are allowed!', 400  # Return a bad request status code
+
+    # Generate the final file name with the extension
+    file_code += file_extension
 
     # Create an instance of the Bucket model
     new_file = Bucket(file_name, file_code, user_id)
@@ -62,22 +69,32 @@ def upload_to_s3():
     db.session.add(new_file)
     db.session.commit()
 
+    # Upload the file to S3
     client.put_object(
         Body=file,
         Bucket=bucket_name,
         Key=file_code
     )
-    
-    return 'File uploaded successfully!', 200
 
-    
+    new_bucket = Bucket.query.filter_by(bucketfile_name=file_name, user_id=user_id).first()
+
+    return {'message': 'File uploaded successfully!', 'id': new_bucket.bucketfile_id}, 200
 
 @bucket_bp.route("/delete/<item_id>", methods=['DELETE'])
 @login_required
 def delete(item_id):
-    file_name = item_id
-    client.delete_object(
-        Bucket=bucket_name,
-        Key=file_name,
-    )
-    return f'File deleted: {file_name}'
+    bucket = Bucket.query.filter_by(bucketfile_id=item_id).first()
+    if bucket:
+        file_name = bucket.bucketfile_code
+        print(file_name)
+        # Delete the file from the bucket or storage system
+        client.delete_object(
+            Bucket=bucket_name,
+            Key=file_name,
+        )
+        # Delete the corresponding MySQL row
+        db.session.delete(bucket)
+        db.session.commit()
+        return f'File deleted: {bucket.bucketfile_name}', 200
+    else:
+        return 404
